# Send sftp packets in parallel